Mod tweaks and perf improvements - zeus:flowers -> zeus:vegetation - Added tall grass to zeus:vegetation - Added new plants - Better Voronoi diagram, output debug images

+++ src/client/gui/DebugDisplay.cpp
@@ -115,7 +115,7 @@ void DebugDisplay::update(sptr<LocalPlayer> player, f64 delta, u32 interpolatedC
 
 	// Textual information
 
-	vec3 playerPos = glm::floor(player->getPos());
+	ivec3 playerPos = glm::floor(player->getPos());
 	vec3 chunkPos = Space::Chunk::world::fromBlock(playerPos);
 	vec3 mapBlockPos = Space::MapBlock::world::fromChunk(chunkPos);
 	vec3 regionPos = Space::Region::world::fromChunk(chunkPos);
